feat: converts loadContracts to react query
